+++ offchain/data/migrations/20230921143147_completion_status/down.sql
@@ -1,0 +1,5 @@
+-- This file should undo anything in `up.sql`
+
+ALTER TABLE "inputs" DROP "status";
+
+DROP TYPE "CompletionStatus";