+++ offchain/data/migrations/20230921143147_completion_status/up.sql
@@ -1,0 +1,14 @@
+-- Your SQL goes here
+
+CREATE TYPE "CompletionStatus" AS ENUM (
+    'Unprocessed',
+    'Accepted',
+    'Rejected',
+    'Exception',
+    'MachineHalted',
+    'CycleLimitExceeded',
+    'TimeLimitExceeded',
+    'PayloadLengthLimitExceeded'
+);
+
+ALTER TABLE "inputs" ADD "status" "CompletionStatus" NOT NULL DEFAULT 'Unprocessed';